Add delete activity endpoint

+++ src/main/java/com/ounal/activity_manager/ActivityRepository.java
@@ -2,6 +2,6 @@ package com.ounal.activity_manager;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ActivityRepository extends JpaRepository<Activity, Integer> {
+public interface ActivityRepository extends JpaRepository<Activity, Long> {
 
 }
